Exception handling - delete

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repositories.UserRepository;
+import com.educandoweb.course.services.exceptions.DatabaseException;
 import com.educandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -15,11 +18,11 @@ public class UserService {
 
 	@Autowired
 	private UserRepository repository;
-	
-	public List<User> findAll(){
+
+	public List<User> findAll() {
 		return repository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = repository.findById(id);
 
@@ -27,23 +30,31 @@ public class UserService {
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id)); // Expressao lambda
 
 	}
-	
+
 //	save() retorna objeto salvo por padrao
- 	public User insert(User obj) {
+	public User insert(User obj) {
 		return repository.save(obj);
 	}
- 	
- 	public void delete(Long id) {
- 		repository.deleteById(id);
- 	}
- 	
- 	public User update(Long id, User obj) {
+
+	public void delete(Long id) {
+		try {
+			if (!repository.existsById(id))
+				throw new ResourceNotFoundException(id);
+			repository.deleteById(id);
+		} catch (EmptyResultDataAccessException e) {
+			throw new ResourceNotFoundException(id);
+		} catch (DataIntegrityViolationException e) {
+			throw new DatabaseException(e.getMessage());
+		}
+	}
+
+	public User update(Long id, User obj) {
 // 		getReferenceById() instancia um usuario monitorado pelo JPA, mas nao vai no banco de dados ainda
- 		User entity = repository.getReferenceById(id);
- 		updateData(entity, obj);
- 		
- 		return repository.save(entity);
- 	}
+		User entity = repository.getReferenceById(id);
+		updateData(entity, obj);
+
+		return repository.save(entity);
+	}
 
 	private void updateData(User entity, User obj) {
 		entity.setName(obj.getName());
